+++ types.h
@@ -45,7 +45,7 @@ public:
 		for(i=0;i<this->n;i++){
 			for(j=0;j<this->m;j++){
 				(*this)(i,j)=
-				(random()-(RAND_MAX/2))*2.0/((float)RAND_MAX)/((float)RANDSCALING);
+				(random()-(RAND_MAX/2))*2.0/((double)RAND_MAX)/((double)RANDSCALING);
 			}
 		}
 	}
@@ -74,7 +74,7 @@ public:
 			el[i]=0;
 		}
 	}
-	Array(float *x,int n){
+	Array(double *x,int n){
 		int i;
 		this->n=n;
 		this->el.resize(n);
@@ -82,7 +82,7 @@ public:
 			for(i=0;i<n;i++){
 				this->el[i]=x[i];
 			}
-//			memcpy(p->el,x,n*sizeof(float));
+//			memcpy(p->el,x,n*sizeof(double));
 		}
 	}
 	~Array(){
@@ -101,7 +101,7 @@ public:
 	}
 	void print(){
 		int i;
-		float *x;
+		double *x;
 		x=this->el;
 		for(i=0;i<this->n;i++){
 			printf("[%i]%.02f\t",i,*x++);
@@ -112,28 +112,28 @@ public:
 		int i;
 		for(i=0;i<this->n;i++){
 			this->el[i]=
-			(random()-(RAND_MAX/2))*2.0/((float)RAND_MAX)/((float)RANDSCALING);
+			(random()-(RAND_MAX/2))*2.0/((double)RAND_MAX)/((double)RANDSCALING);
 		}
 	}
 };
 
 class Layer{
 public:
-	Matrix<float> *M;
-	Array<float> *out;
-	Array<float> *deriv;
-	Array<float> *delta;
+	Matrix<double> *M;
+	Array<double> *out;
+	Array<double> *deriv;
+	Array<double> *delta;
 	Layer(int n,int m){
-		out=new Array<float>(n);
-		deriv=new Array<float>(n);
-		delta=new Array<float>(n);
-		M=new Matrix<float>(n,m);
+		out=new Array<double>(n);
+		deriv=new Array<double>(n);
+		delta=new Array<double>(n);
+		M=new Matrix<double>(n,m);
 	}
 	~Layer(){
 	}
-	Array<float> *forward(const Array<float> *x){
+	Array<double> *forward(const Array<double> *x){
 		int i,j;
-		float a,tmp;
+		double a,tmp;
 		for(j=0;j<M->n;j++){
 			a=0;
 			for(i=0;i<M->m;i++){
@@ -147,16 +147,16 @@ public:
 		}
 		return(this->out);
 	}
-	void outputDelta(const Array<float> *error){
+	void outputDelta(const Array<double> *error){
 		int j;
 		for(j=0;j<error->n;j++){
 			(*this->delta)(j)=(*this->deriv)(j)*(*error)(j);
 			//delta->el[j]=error->el[j];
 		}
 	}
-	void upDelta(const Matrix<float> *W,const Array<float> *delta2){
+	void upDelta(const Matrix<double> *W,const Array<double> *delta2){
 		int j,k;
-		float sum;
+		double sum;
 		//for(j=0;j<this->deriv->n;j++){
 		for(j=0;j<W->m;j++){
 			sum=0;
@@ -169,7 +169,7 @@ public:
 			(*this->delta)(j)=(*this->deriv)(j)*sum;
 		}
 	}
-	void updateWeights(const Array<float> *input){
+	void updateWeights(const Array<double> *input){
 		int i,j;
 		for(i=0;i<this->M->n;i++){
 			for(j=0;j<this->M->m;j++){
@@ -186,8 +186,8 @@ class Net{
 public:
 	Layer **L;
 	int n;
-	Array<float> *error;
-	Array<float> *answer;
+	Array<double> *error;
+	Array<double> *answer;
 	Net(int n=0){
 		this->n=n;
 		L=(Layer **)malloc(n*sizeof(Layer *));
@@ -201,16 +201,16 @@ public:
 	void insertLayer(int i,int n,int m){
 		L[i]=new Layer(n,m);
 		if(i==(this->n-1)){
-			error=new Array<float>(n);
-			answer=new Array<float>(n);
+			error=new Array<double>(n);
+			answer=new Array<double>(n);
 		}
 	}
-	void forward(const Array<float> *x){
+	void forward(const Array<double> *x){
 		//int i;
 		L[0]->forward(x);
 		L[1]->forward(L[0]->out);
 /*
-		Array<float> *t=x;
+		Array<double> *t=x;
 		for(i=0;i<this->n;i++){
 			t=L[i]->forward(t);
 	//		MatMul(*N->L[i]->M,*N->L[i]->in,*N->L[i]->out,*N->L[i]->deriv);
@@ -218,7 +218,7 @@ public:
 */
 		this->answer=L[1]->out;
 	}
-	void backward(const Array<float> *input){
+	void backward(const Array<double> *input){
 		//int i;
 		L[1]->outputDelta(this->error);
 		L[0]->upDelta(L[1]->M,L[1]->delta);
@@ -246,13 +246,13 @@ public:
 			L[i]->M->print();
 		}
 	}
-	Array<float> *train(const Array<float> *x,const Array<float> *y){
+	Array<double> *train(const Array<double> *x,const Array<double> *y){
 		this->forward(x);
 		this->upError(y);
 		this->backward(x);
 		return(this->error);
 	}
-	void upError(const Array<float> *yTarget){
+	void upError(const Array<double> *yTarget){
 		int i;
 		for(i=0;i<yTarget->n;i++){
 			(*this->error)(i)=(*this->answer)(i)-(*yTarget)(i);
@@ -262,10 +262,10 @@ public:
 
 #endif
 /*
-float nnTotalError(const Array<float> *y0,const Array<float> *y){
+double nnTotalError(const Array<double> *y0,const Array<double> *y){
 	int i;
 	int n=y0->n;
-	float ret=0;
+	double ret=0;
 	for(i=0;i<n;i++){
 		ret+=fabs(y0->el[i]-y->el[i]);
 		ret*=ret;
